requirements.txt
telethon==1.16.4
keyboard==0.13.5
colorama==0.4.3
pyfiglet==0.8.post1
requests==2.25.1
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
python-dotenv==0.18.0
pyqt5